requirements.txt
aiomysql==0.2.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
click==8.1.8
colorlog==6.9.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-decouple==3.8
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0

